refactor: ImageModel 프로퍼티명 url -> uri로 변경 및 패키지 savelog -> model로 변경
