+++ app/src/main/java/com/treasurehunt/ui/model/ImageModel.kt
@@ -1,0 +1,5 @@
+package com.treasurehunt.ui.model
+
+data class ImageModel(
+    val uri: String,
+)
