+++ app/src/main/java/com/treasurehunt/ui/savelog/ImageClickListener.kt
@@ -1,5 +1,7 @@
 package com.treasurehunt.ui.savelog
 
+import com.treasurehunt.ui.model.ImageModel
+
 fun interface ImageClickListener {
     fun onClick(imageModel: ImageModel)
 }
