--- app/src/main/java/com/treasurehunt/ui/savelog/ImageModel.kt
@@ -1,5 +1,0 @@
-package com.treasurehunt.ui.savelog
-
-data class ImageModel(
-    val url: String,
-)
